fix(train): gates for the nursery RegExp, the page-class table, and the LIFO handle stack
Four gate failures the assembled tree produced, and the rooting bug the
suite caught:

- The runtime handle stack is strictly LIFO (`Drop` truncates to the
  scope's base), so rooting into an OUTER scope while an inner one is
  live has the inner scope's drop discard the handle. #9869's
  `visited.push(&scope, ..)` sat inside #9864's per-level scope and hit
  "runtime handle used after its scope was dropped". The per-level scope
  now closes before the push. Caught by
  gc::tests::rooted_for_in::for_in_grown_result_and_receiver_survive_prototype_collection.

- shape_descriptor_census asserted `gc_malloc(.. GC_TYPE_REGEXP)` at
  `js_regexp_new`; #9845 deliberately moves that birth to the nursery, so
  the assertion now accepts either allocator. What it checks is unchanged
  and is the point: RegExp is born with its OWN GcHeader kind, never as a
  generic object something later re-identifies by payload magic. Verified
  the updated gate still fails when the birth kind is blunted.

- #9853's page-class table pushed arena/page_meta.rs to 2559 lines. Split
  into page_meta/{mod,page_class,tests}.rs; the page-class tests move next
  to their subject. Both feature configurations build.

- That split also stranded six frontier entries in
  gc_runtime_root_holders.json on the old path, and the PASS1_MARKED
  census pin needed its re-audit for #9860's and #9845's gc/mod.rs
  re-export additions before the hash could move.
